% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agriculture.R
\docType{data}
\name{Agriculture}
\alias{Agriculture}
\title{Agriculture}
\format{A data frame with 9 rows and 4 variables. The variables
are as follows:
\describe{
  \item{Y}{The yield values (Dependent variables values)}
  \item{X1}{The fertilizer values (The first independent variables values)} 
  \item{X2}{The irrigation values (The second independent variables values)}
  \item{D}{The experience values (The third independent values)}
}}
\source{
The data set is used in the book entitled
Multivariate Statistical Methods with R Applications (R Uygulamalari ile Cok Degiskenli Istatistiksel Yontemler), H.Bulut (2018).
}
\usage{
Agriculture
}
\description{
The data set is used in Sample 8.3. The data set consists of
4 variables and 9 observations. This data is imaginary. Moreover the first three variables 
of this data are used in Sample 8.2.
}
\keyword{datasets}
