% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\docType{data}
\name{ecodata}
\alias{ecodata}
\title{Economic Indicators of Cities in Turkey}
\format{A data frame with 81 rows and 4 variables. The variables
are as follows:
\describe{
  \item{Cities}{The names of cities in Turkey}
  \item{X1}{The employment rate in cities} 
  \item{X2}{The unemployment rate in cities} 
  \item{X3}{The satisfaction rate from job} 
}}
\source{
The data set is taken from Turkish Statistical Institute (TUIK) 
and used in the book entitled Multivariate Statistical Methods with R Applications (R Uygulamalari ile Cok Degiskenli Istatistiksel Yontemler), H.Bulut (2018).
}
\usage{
ecodata
}
\description{
The data set is used in Sample 14.2 and given in Table 14.2. 
The data set consists of 4 variables and 81 observations. This data contains 
some economic indicators of cities in Turkey.
}
\keyword{datasets}
