% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SER.R
\docType{data}
\name{SER}
\alias{SER}
\title{Sex, Education and Reading Data}
\format{
A data frame with 21 rows and 3 variables. The variables
are as follows:
\describe{
  \item{Sex}{The sex of observations}
  \item{Education}{The education level of observations} 
  \item{Reading}{The reading case of observations} 
}
}
\source{
The data set is used in the book entitled
Multivariate Statistical Methods with R Applications (R Uygulamalari ile Cok Degiskenli Istatistiksel Yontemler), H.Bulut (2018).
}
\usage{
SER
}
\description{
The data set is used in Sample 15.2 and given in Table 15.1. 
The data set consists of 3 variables and 21 observations. This data is imaginary.
}
\keyword{datasets}
