% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim1.R
\docType{data}
\name{Sim1}
\alias{Sim1}
\title{Simulated Data 1}
\format{
A data frame with 10 rows and 3 variables.
}
\source{
The data set is used in the book entitled
Multivariate Statistical Methods with R Applications (R Uygulamalari ile Cok Degiskenli Istatistiksel Yontemler), H.Bulut (2018).
}
\usage{
Sim1
}
\description{
The data set is used in Sample 3.1. The data set consists of
3 variables and 10 observations. This data is imaginary.
}
\keyword{datasets}
