% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim2.R
\docType{data}
\name{Sim2}
\alias{Sim2}
\title{Simulated Data 2}
\format{A data frame with 10 rows and 2 variables.}
\source{
The data set is used in the book entitled
Multivariate Statistical Methods with R Applications (R Uygulamalari ile Cok Degiskenli Istatistiksel Yontemler), H.Bulut (2018).
}
\usage{
Sim2
}
\description{
The data set is used in Sample 3.4. The data set consists of
2 variables and 10 observations. This data is simulated from multivariate 
normal distribition with mean vector (5,10) and covariance matrix [9,10,10,25].
}
\keyword{datasets}
