% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssl.R
\docType{data}
\name{ssl}
\alias{ssl}
\title{Security and Social Life Indicators of Cities in Turkey}
\format{
A data frame with 81 rows and 9 variables. The variables
are as follows:
\describe{
  \item{Cities}{The names of cities in Turkey}
  \item{X1}{The murder rate in cities} 
  \item{X2}{The number of traffic accidents in cities} 
  \item{X3}{The rate of people who feel safe themselves when walking alone at night} 
  \item{X4}{The rate of satisfaction from public security services} 
  \item{Y1}{Cinema and theater audience} 
  \item{Y2}{Shopping center area per thousand people} 
  \item{Y3}{Satisfaction with social relations} 
  \item{Y4}{Satisfaction with social life} 
}
}
\source{
The data set is taken from Turkish Statistical Institute (TUIK) 
and used in the book entitled Multivariate Statistical Methods with R Applications (R Uygulamalari ile Cok Degiskenli Istatistiksel Yontemler), H.Bulut (2018).
}
\usage{
ssl
}
\description{
The data set is used in Sample 13.2 and given in Table 13.1. 
The data set consists of 9 variables and 81 observations. This data contains 
security and social life indicators of cities in Turkey.
}
\keyword{datasets}
